// db/types.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType, Insertable, Updateable } from "kysely";

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Prizes {
  created_at: Generated<Timestamp>;
  document_id: Generated<string>;
  id: Generated<number>;
  live: boolean | null;
  published_at: Timestamp | null;
  slug: string;
  subtitle: string | null;
  ticket_price: Generated<number>;
  tickets_sold: Generated<number>;
  tickets_total: number;
  title: string;
  updated_at: Generated<Timestamp>;
}

export interface Users {
  clerk_id: string;
  created_at: Generated<Timestamp>;
  email: string;
  first_name: string | null;
  id: Generated<number>;
  image_url: string | null;
  last_name: string | null;
  updated_at: Generated<Timestamp>;
  username: string | null;
}

export type User = Selectable<Users>
export type NewUser = Insertable<Users>
export type UserUpdate = Updateable<Users>

export interface DB {
  prizes: Prizes;
  users: Users;
}